requirements.txt
beautifulsoup4==4.12.3
Django==5.0.3
python-dotenv==1.0.1

djangorestframework~=3.14.0
psutil~=5.9.8